app/requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests==2.26.0
gunicorn==20.1.0
pytest==6.2.4
Jinja2==3.0.1
protobuf==3.20.3
grpcio==1.54.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-semantic-conventions==0.42b0
prometheus-flask-exporter==0.22.3
python-logging-loki==0.3.1
prometheus-client==0.17.1 